+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+* [PR-25](https://github.com/ITK-Leantime/leantime-timetable/pull/25)
+  * Added quick-add ticket to timetable
+  * Added possibility to add a ticket via API
+
 * [PR-24](https://github.com/ITK-Leantime/leantime-timetable/pull/24)
   * Hide vertical scrollbar on description field on windows
   * Remove concept of filtering tickets